Changed Semantic Props back to CSS class from Web Component.
