+++ src/font.css
@@ -1,6 +1,6 @@
 /* Font: Custom properties for styling fonts. */
 
-semantic-props {
+:root, .--semantic {
 	/* Font Families: */
 	--font: Seravek, 'Gill Sans Nova', Ubuntu, Calibri, 'DejaVu Sans', source-sans-pro, sans-serif;
 	--monospace-font: ui-monospace, 'Cascadia Code', 'Source Code Pro', Menlo, Consolas, 'DejaVu Sans Mono', monospace;
