+++ src/layer.css
@@ -1,6 +1,6 @@
 /* Layer: Custom properties for z-index positions. */
 
-semantic-props {
+.--semantic {
 	/* Layer Positions: */
 	--bottom-layer: -1000;
 	--relative-layer: 1000;
