+++ src/safe.css
@@ -1,6 +1,6 @@
 /* Safe: Custom properties for display safe areas. */
 
-semantic-props {
+:root, .--semantic {
 	/* Safe Areas: */
 	--safe-top: env(safe-area-inset-top);
 	--safe-right: env(safe-area-inset-right);
